Update main.py
The user doesn't see the correct number in the case, that they almost got it and can try again so that they don't know the number. Also fixxed spelling mistakes and made the while loop break in the case of a number being guessed which is out of the guessing range.